fix!: change ids of type `ulong` to type `long`
Chunk id and directory id both seem to be stored as a signed integer in
`Riot Client.db` and I think it's safe to assume that file and bundle id
should be signed as well.

Id for `Language` is different because it's incremental (file languages
are also stored as a bit mask), so it's *probably* unsigned.